+ coordinator layout + load more

+++ app/src/main/java/com/app/hcmut/movie/model/Movie.kt
@@ -33,7 +33,7 @@ class Movie {
     @SerializedName("release_date")
     var releaseDate: String? = null
 
-    var revenue: Int? = null
+    var revenue: Long? = null
 
     var runtime: Int? = null
